+++ src/behaviours/draggable.js
@@ -9,7 +9,7 @@ import { filter } from 'lodash';
 import DraggablePreview from '../utils/DraggablePreview';
 import { actionCreators as draggableActions } from '../ducks/modules/draggable';
 import { actionCreators as droppableActions } from '../ducks/modules/droppable';
-import { PreventGhostClick } from '.';
+import PreventGhostClick from '../utils/PreventGhostClick';
 
 function isTouch(event) {
   if (typeof TouchEvent !== 'undefined' && event instanceof TouchEvent) {
