+++ src/behaviours/index.js
@@ -1,6 +1,5 @@
 export { default as draggable } from './draggable';
 export { default as droppable } from './droppable';
-export { default as PreventGhostClick } from './PreventGhostClick';
 export { default as selectable } from './selectable';
 export { default as scrollable } from './scrollable';
 export { default as autoInitialisedForm } from './autoInitialisedForm';
